a little more conservative "deny" provisions - in case of multiple codings
This change should be a no-op. Only in the case where within the same Consent two provisions with identical codes show up (one with deny and one with permit) would the deny now count (regardless of order) (therefore the check on existing code2consent content for coding.code). Also if within a codeable concept different codings are listed - each should show up in case of a deny (therefore no break).